add general order template and fix checkstyle
