+++ src/main/java/uk/gov/hmcts/reform/civil/model/CaseData.java
@@ -28,7 +28,6 @@ import uk.gov.hmcts.reform.civil.model.genapplication.GAUrgencyRequirement;
 import uk.gov.hmcts.reform.civil.model.genapplication.GeneralApplication;
 import uk.gov.hmcts.reform.civil.model.genapplication.GeneralApplicationsDetails;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
