+++ src/main/java/uk/gov/hmcts/reform/civil/model/docmosis/dismissalorder/DismissalOrder.java
@@ -1,0 +1,33 @@
+package uk.gov.hmcts.reform.civil.model.docmosis.dismissalorder;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateSerializer;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import uk.gov.hmcts.reform.civil.model.common.MappableObject;
+
+import java.time.LocalDate;
+
+@Getter
+@Builder
+@AllArgsConstructor
+@EqualsAndHashCode
+public class DismissalOrder implements MappableObject {
+
+    private final String claimNumber;
+    private final String claimantName;
+    private final String defendantName;
+    private final String applicationType;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
+    @JsonSerialize(using = LocalDateSerializer.class)
+    private final LocalDate submittedOn;
+    private final String applicantName;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
+    @JsonSerialize(using = LocalDateSerializer.class)
+    private final LocalDate applicationDate;
+    private final String dismissalOrder;
+    private final String reasonForDecision;
+}
