+++ src/main/java/uk/gov/hmcts/reform/civil/model/documents/DocumentType.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.reform.civil.model.documents;
 
 public enum DocumentType {
 
+    GENERAL_ORDER,
     DIRECTION_ORDER,
     DISMISSAL_ORDER,
     REQUEST_FOR_INFORMATION,
