+++ src/main/java/uk/gov/hmcts/reform/civil/service/docmosis/hearingorder/HearingOrderGenerator.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.reform.civil.service.docmosis.writtenRepresentationSequentialOrder;
+package uk.gov.hmcts.reform.civil.service.docmosis.hearingorder;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.civil.enums.dq.GeneralApplicationTypes;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.docmosis.DocmosisDocument;
-import uk.gov.hmcts.reform.civil.model.docmosis.writtenRepresentationSequentialOrder.WrittenRepSequentialOrder;
+import uk.gov.hmcts.reform.civil.model.docmosis.hearingorder.HearingOrder;
 import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
 import uk.gov.hmcts.reform.civil.model.documents.DocumentType;
 import uk.gov.hmcts.reform.civil.model.documents.PDF;
@@ -19,17 +19,17 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static uk.gov.hmcts.reform.civil.service.docmosis.DocmosisTemplates.WRITTEN_REP_SEQUENTIAL;
+import static uk.gov.hmcts.reform.civil.service.docmosis.DocmosisTemplates.HEARING_ORDER;
 
 @Service
 @RequiredArgsConstructor
-public class WrittenRepresentationSequentailOrderGenerator implements TemplateDataGenerator<WrittenRepSequentialOrder> {
+public class HearingOrderGenerator implements TemplateDataGenerator<HearingOrder> {
 
     private final DocumentManagementService documentManagementService;
     private final DocumentGeneratorService documentGeneratorService;
 
     public CaseDocument generate(CaseData caseData, String authorisation) {
-        WrittenRepSequentialOrder templateData = getTemplateData(caseData);
+        HearingOrder templateData = getTemplateData(caseData);
 
         DocmosisTemplates docmosisTemplate = getDocmosisTemplate(caseData);
 
@@ -41,7 +41,7 @@ public class WrittenRepresentationSequentailOrderGenerator implements TemplateDa
         return documentManagementService.uploadDocument(
             authorisation,
             new PDF(getFileName(docmosisTemplate, caseData), docmosisDocument.getBytes(),
-                    DocumentType.WRITTEN_REP_SEQUENTIAL)
+                    DocumentType.HEARING_ORDER)
         );
     }
 
@@ -50,11 +50,7 @@ public class WrittenRepresentationSequentailOrderGenerator implements TemplateDa
     }
 
     @Override
-    public WrittenRepSequentialOrder getTemplateData(CaseData caseData) {
-        List<GeneralApplicationTypes> types = caseData.getGeneralAppType().getTypes();
-        String collect = types.stream()
-            .map(GeneralApplicationTypes::getDisplayedValue).collect(Collectors.joining(", "));
-
+    public HearingOrder getTemplateData(CaseData caseData) {
         List<String> claimantNames = new ArrayList<>();
         claimantNames.add(caseData.getClaimant1PartyName());
         if (caseData.getClaimant2PartyName() != null) {
@@ -69,24 +65,26 @@ public class WrittenRepresentationSequentailOrderGenerator implements TemplateDa
         }
         String defendantName = String.join(", ", defendentNames);
 
-        WrittenRepSequentialOrder.WrittenRepSequentialOrderBuilder writtenRepSequentialOrderBuilder =
-            WrittenRepSequentialOrder.builder()
+        List<GeneralApplicationTypes> types = caseData.getGeneralAppType().getTypes();
+        String collect = types.stream()
+            .map(GeneralApplicationTypes::getDisplayedValue).collect(Collectors.joining(", "));
+
+        HearingOrder.HearingOrderBuilder hearingOrderBuilder =
+            HearingOrder.builder()
                 .claimNumber(caseData.getCcdCaseReference().toString())
                 .applicationType(collect)
                 .claimantName(claimantName)
                 .defendantName(defendantName)
                 .applicantName(caseData.getApplicantPartyName())
                 .applicationDate(caseData.getCreatedDate().toLocalDate())
-                .uploadDeadlineDate(caseData.getJudicialDecisionMakeAnOrderForWrittenRepresentations()
-                                        .getWrittenSequentailRepresentationsBy())
-                .responseDeadlineDate(caseData.getJudicialDecisionMakeAnOrderForWrittenRepresentations()
-                                          .getSequentialApplicantMustRespondWithin())
+                .hearingLocation("TEST")
+                .estimatedHearingLength(caseData.getJudicialGeneralOrderHearingEstimationTimeText())
                 .submittedOn(LocalDate.now());
 
-        return writtenRepSequentialOrderBuilder.build();
+        return hearingOrderBuilder.build();
     }
 
     private DocmosisTemplates getDocmosisTemplate(CaseData caseData) {
-        return WRITTEN_REP_SEQUENTIAL;
+        return HEARING_ORDER;
     }
 }
