+++ src/main/java/uk/gov/hmcts/reform/civil/service/docmosis/writtenrepresentationsequentialorder/WrittenRepresentationSequentailOrderGenerator.java
@@ -1,11 +1,11 @@
-package uk.gov.hmcts.reform.civil.service.docmosis.requestmoreinformation;
+package uk.gov.hmcts.reform.civil.service.docmosis.writtenrepresentationsequentialorder;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.civil.enums.dq.GeneralApplicationTypes;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.docmosis.DocmosisDocument;
-import uk.gov.hmcts.reform.civil.model.docmosis.requestforinformation.RequestForInformation;
+import uk.gov.hmcts.reform.civil.model.docmosis.writtenrepresentationsequentialorder.WrittenRepSequentialOrder;
 import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
 import uk.gov.hmcts.reform.civil.model.documents.DocumentType;
 import uk.gov.hmcts.reform.civil.model.documents.PDF;
@@ -19,17 +19,17 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static uk.gov.hmcts.reform.civil.service.docmosis.DocmosisTemplates.REQUEST_FOR_INFORMATION;
+import static uk.gov.hmcts.reform.civil.service.docmosis.DocmosisTemplates.WRITTEN_REP_SEQUENTIAL;
 
 @Service
 @RequiredArgsConstructor
-public class RequestForInformationGenerator implements TemplateDataGenerator<RequestForInformation> {
+public class WrittenRepresentationSequentailOrderGenerator implements TemplateDataGenerator<WrittenRepSequentialOrder> {
 
     private final DocumentManagementService documentManagementService;
     private final DocumentGeneratorService documentGeneratorService;
 
     public CaseDocument generate(CaseData caseData, String authorisation) {
-        RequestForInformation templateData = getTemplateData(caseData);
+        WrittenRepSequentialOrder templateData = getTemplateData(caseData);
 
         DocmosisTemplates docmosisTemplate = getDocmosisTemplate(caseData);
 
@@ -41,7 +41,7 @@ public class RequestForInformationGenerator implements TemplateDataGenerator<Req
         return documentManagementService.uploadDocument(
             authorisation,
             new PDF(getFileName(docmosisTemplate, caseData), docmosisDocument.getBytes(),
-                    DocumentType.REQUEST_FOR_INFORMATION)
+                    DocumentType.WRITTEN_REP_SEQUENTIAL)
         );
     }
 
@@ -50,7 +50,7 @@ public class RequestForInformationGenerator implements TemplateDataGenerator<Req
     }
 
     @Override
-    public RequestForInformation getTemplateData(CaseData caseData) {
+    public WrittenRepSequentialOrder getTemplateData(CaseData caseData) {
         List<String> claimantNames = new ArrayList<>();
         claimantNames.add(caseData.getClaimant1PartyName());
         if (caseData.getClaimant2PartyName() != null) {
@@ -69,20 +69,24 @@ public class RequestForInformationGenerator implements TemplateDataGenerator<Req
         String collect = types.stream()
             .map(GeneralApplicationTypes::getDisplayedValue).collect(Collectors.joining(", "));
 
-        RequestForInformation.RequestForInformationBuilder requestForInformationBuilderBuilder =
-            RequestForInformation.builder()
+        WrittenRepSequentialOrder.WrittenRepSequentialOrderBuilder writtenRepSequentialOrderBuilder =
+            WrittenRepSequentialOrder.builder()
                 .claimNumber(caseData.getCcdCaseReference().toString())
                 .applicationType(collect)
                 .claimantName(claimantName)
                 .defendantName(defendantName)
-                .judgeComments(caseData.getJudicialDecisionRequestMoreInfo().getJudgeRequestMoreInfoText())
-                .submittedOn(LocalDate.now().plusDays(14))
-                .dateBy(caseData.getJudicialDecisionRequestMoreInfo().getJudgeRequestMoreInfoByDate());
+                .applicantName(caseData.getApplicantPartyName())
+                .applicationDate(caseData.getCreatedDate().toLocalDate())
+                .uploadDeadlineDate(caseData.getJudicialDecisionMakeAnOrderForWrittenRepresentations()
+                                        .getWrittenSequentailRepresentationsBy())
+                .responseDeadlineDate(caseData.getJudicialDecisionMakeAnOrderForWrittenRepresentations()
+                                          .getSequentialApplicantMustRespondWithin())
+                .submittedOn(LocalDate.now());
 
-        return requestForInformationBuilderBuilder.build();
+        return writtenRepSequentialOrderBuilder.build();
     }
 
     private DocmosisTemplates getDocmosisTemplate(CaseData caseData) {
-        return REQUEST_FOR_INFORMATION;
+        return WRITTEN_REP_SEQUENTIAL;
     }
 }
